+++ tests/test_time.cu
@@ -46,7 +46,7 @@ bool run_test(int B, int T, int L, int A, int num_threads) {
     std::vector<int> label_lengths;
     for (const auto& l : labels) {
         flat_labels.insert(flat_labels.end(), l.begin(), l.end());
-        label_lengths.push_back(l.size());
+        label_lengths.push_back((int)l.size());
     }
 
     std::vector<float> costs(B);
@@ -104,7 +104,7 @@ bool run_test(int B, int T, int L, int A, int num_threads) {
 
         cudaFree(rnnt_gpu_workspace);
         elapsed = end - start;
-        time.push_back(elapsed.count() * 1000);
+        time.push_back((float)(elapsed.count() * 1000));
         std::cout << "compute_rnnt_loss elapsed time: " << elapsed.count() * 1000 << " ms\n";
     }
 
@@ -127,7 +127,7 @@ bool run_test(int B, int T, int L, int A, int num_threads) {
 
     std::cout << "average 10 time cost: " << sum << " ms variance: " << std << std::endl;
 
-    float cost = std::accumulate(costs.begin(), costs.end(), 0.);
+    float cost = std::accumulate(costs.begin(), costs.end(), 0.0f);
 
     free(acts);
     return true;
